refactor: move extension methods into core
Extension methods for View and Fragment will be shared amongst over modules, so move them over to
the core module for reusability
