+++ core/build.gradle
@@ -49,4 +49,6 @@ dependencies {
 
     testImplementation 'io.mockk:mockk:1.9.3'
     testImplementation 'junit:junit:4.12'
+
+    androidTestImplementation 'androidx.test.ext:junit:1.1.1'
 }
