+++ core/src/androidTest/java/com/chesire/malime/core/extensions/ViewExtensionsTests.kt
@@ -1,4 +1,4 @@
-package com.chesire.malime.extensions
+package com.chesire.malime.core.extensions
 
 import android.view.View
 import androidx.test.ext.junit.runners.AndroidJUnit4
